2022.03.17 (목) 김희은 수정 [엔티티 변경사항] item: 제조사컬럼 nullable 로 변경 bomItemDetail: bom수량 컬럼 타입 float 로 변경

+++ src/main/java/com/mes/mesBackend/dto/request/BomItemRequest.java
@@ -26,7 +26,7 @@ public class BomItemRequest {
 
     @Schema(description = "수량")
     @NotNull(message = NOT_NULL)
-    int amount;
+    float amount;
 
     @Schema(description = "거래처 id")
     @Min(value = ONE_VALUE, message = NOT_ZERO)
